Fix breakage on tooltips
It turns out that window-configuration-change-hook gets called a lot,
in a lot of random circumstances. One of those is the creation of
tooltips. However, for some reason, during the creation of a tooltip,
trying to call (face-attr-construct 'ethan-wspace-face) dies with

Error while displaying tooltip: (error Invalid face ethan-wspace-face)

So now we try to detect tooltip creation and avoid doing anything when
it happens.