Changing major mode doesn't create more overlays
Fix a bug where changing major modes, i.e. by reverting a buffer,
would lose the value of the old overlay, and then create a new one,
leading to an accumulation of no-nl-eof overlays.  To get around this,
we set the permanent-local property on the overlay variable, which
prevents it from being killed by kill-all-local-variables.